Fixes and improvements
Improved the DefaultPanelList class to accept variable color fills, and removed the static color variables in favor of using the setColor() method to do so.

As a result, changing the default panelList's color (e.g. for theming) will require the use of a custom PanelListFactory which keeps track of all the panelLists created via that factory, and updating the color for all of them at once, on demand; this method will require extra work on the part of the developers, but is more flexible in that multiple Docker instances and their containers can have varying theme colors for the same (default) panelList implementation.