Added delegates, tightened DefaultPanelList class
Added the DockHelperDelegate, PanelDelegate and ResizerDelegate classes. Also, tightened the DefaultPanelList class to remove the IDockTarget interface (as there is no guarantee that it can accept all NativeDragEvents, since it has no knowledge of drag clipboard formats ahead-of-time) - this feature is left up to the developer, or deferred for a future release.