New IContainer decorators, pair classes, and fixes
Added extra pair classes (RandomPair and CyclicPair) and added the two decorators, ContainerBorder, and ContainerMask, for adding a border around an IContainer instance, and to mask the contents of an IContainer instance, respectively. To add these decorators, override the default IContainerFactory implementation's createContainer() method to return an IContainer instance with these decorators added to the default IContainer implementation (or any implementation of choice.)

Fixed some extra bugs and added the isAntiComplementary() and isAntiComplementaryString() functions to PanelContainerSide as well, to detect whether a side is neither the same, nor complementary.